remove the obsolete utils.fitsutil and utils.dspec
