--- suncasa/shellScript/pipeline_allday_fits.csh
@@ -1,5 +1,0 @@
-#! /bin/tcsh -f
-
-source /home/user/.cshrc
-# Run the allday fits pipeline
-/common/anaconda2/bin/python /common/python/current/pipeline_allday_fits.py $argv
